+++ packages/inputs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @under-control/inputs
 
+## 1.3.0
+
+### Minor Changes
+
+- Add caching support for onChange
+
+### Patch Changes
+
+- Updated dependencies
+  - @under-control/core@1.3.0
+
 ## 1.2.1
 
 ### Patch Changes